@@ -22,10 +22,10 @@
   },
   "dependencies": {
     "history": "^4.7.2",
-    "react": "16.0.0",
-    "react-dom": "16.0.1",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-redux": "^5.0.6",
-    "react-router-dom": "^4.2.2",
+    "react-router-dom": "^5.1.2",
     "redux": "^3.7.2",
     "redux-form": "^8.3.0",
     "redux-logger": "^3.0.6",
